Xong chức năng xóa thiết bị

+++ src/main/java/com/example/QLThanhVien/Controller/ThietBiController.java
@@ -2,7 +2,9 @@ package com.example.QLThanhVien.Controller;
 
 
 import com.example.QLThanhVien.Enity.ThietBiEntity;
+import com.example.QLThanhVien.Enity.ThongTinSuDungEntity;
 import com.example.QLThanhVien.Repository.ThietBiRepository;
+import com.example.QLThanhVien.Repository.ThongTinSuDungRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.Banner;
 import org.springframework.http.ResponseEntity;
@@ -19,6 +21,9 @@ public class ThietBiController {
 
     @Autowired
     private ThietBiRepository thietBiRepository;
+
+    @Autowired
+    private ThongTinSuDungRepository thongTinSuDungRepository;
 
 
 
@@ -50,21 +55,59 @@ public class ThietBiController {
 
     @DeleteMapping("/ThietBi.html")
     public void deleteDevice(@RequestBody List<Integer> list_id){
+
         for (Integer id : list_id){
 
+            // Danh sach thong tin su dung
+            Iterable<ThongTinSuDungEntity> list = thongTinSuDungRepository.findAll();
 
+            // Danh sách thiet bi
             Optional<ThietBiEntity> temp = thietBiRepository.findById(id);
 
             if (temp.isPresent()) {
+
+                // Ep kieu thanh kieu thietbiEntity
                 ThietBiEntity thietBi = temp.get();
 
-                thietBiRepository.delete(thietBi);
-            } else {
-                System.out.println("Khong tim ra " + id);
-            }
+                // Kiem tra xem thiet bi co dand duoc muon hay dat cho khong
+                ThongTinSuDungEntity thongTinSuDung = CheckMuonvaDatCho((List<ThongTinSuDungEntity>) list,thietBi.getMaTB());
 
+
+                if (thongTinSuDung != null){
+                    System.out.println(thongTinSuDung.getMaTB().getMaTB());
+                }
+                else{
+                    XoaHetThongTinSuDung((List<ThongTinSuDungEntity>) list,id);
+                    thietBiRepository.delete(thietBi);
+                }
+
+
+            }
         }
 
+    }
+
+    public ThongTinSuDungEntity CheckMuonvaDatCho (List<ThongTinSuDungEntity> list,int idThietBi){
+        for (ThongTinSuDungEntity temp : list){
+            if (temp.getMaTB() != null) {
+                if (temp.getMaTB().getMaTB() == idThietBi) {
+                    if (temp.getTGTra() != null || temp.getTGDatCho() != null) {
+                        return temp;
+                    }
+                }
+            }
+        }
+        return null;
+    }
+
+    public void XoaHetThongTinSuDung(List<ThongTinSuDungEntity> list,int idThietBi){
+        for (ThongTinSuDungEntity temp : list){
+            if (temp.getMaTB() != null) {
+                if (temp.getMaTB().getMaTB() == idThietBi) {
+                    thongTinSuDungRepository.delete(temp);
+                }
+            }
+        }
     }
